+++ shuup_tests/admin/test_view_settings.py
@@ -8,13 +8,16 @@ import json
 
 import pytest
 from django.utils.http import urlencode
+
 from shuup import configuration
 from shuup.admin.modules.products.views import ProductListView
 from shuup.admin.modules.settings.views import ListSettingsView
+from shuup.testing.factories import get_default_shop
 
 
 @pytest.mark.django_db
 def test_view_default_columns(rf):
+    shop = get_default_shop()
 
     view = ProductListView.as_view()
 
@@ -30,35 +33,34 @@ def test_view_default_columns(rf):
     column_names = [c.id for c in sorted(listview.columns, key=lambda x: x.id)]
     default_column_names = [c.id for c in sorted(listview.default_columns, key=lambda x: x.id)]
     assert column_names == default_column_names
-
-    assert configuration.get(None, "view_configuration_product_name")  # name is configured
+    assert configuration.get(None, "view_configuration_shopproduct_name")  # name is configured
     assert listview.settings.view_configured()
-    assert listview.settings.get_settings_key("name") == "view_configuration_product_name"  # we are attached to product view
+    assert listview.settings.get_settings_key("name") == "view_configuration_shopproduct_name"  # we are attached to product view
 
     settings_view = ListSettingsView.as_view()
-    view_data = {"model": "Product", "module": "shuup.core.models", "return_url": "product"}
+    view_data = {"model": "ShopProduct", "module": "shuup.core.models", "return_url": "shop_product"}
     request = rf.get("/", view_data)
     response = settings_view(request)
     assert 200 <= response.status_code < 300
 
     # Change configuration by posting form
-    request = rf.post("/?" + urlencode(view_data), {"view_configuration_product_name": False})
+    request = rf.post("/?" + urlencode(view_data), {"view_configuration_shopproduct_name": False})
     response = settings_view(request)
     assert response.status_code == 302
 
-    assert listview.settings.get_config("name") == configuration.get(None, "view_configuration_product_name")
-    assert not configuration.get(None, "view_configuration_product_name").get("active")
+    assert listview.settings.get_config("name") == configuration.get(None, "view_configuration_shopproduct_name")
+    assert not configuration.get(None, "view_configuration_shopproduct_name").get("active")
 
 
 @pytest.mark.django_db
 def test_view_saved_columns(rf):
-    visible_fields = sorted(["id", "name", "select"])
-    configuration.set(None, "view_configuration_product_saved", True)
+    shop = get_default_shop()
+    visible_fields = sorted(["shopproduct_id", "name", "select"])
+    configuration.set(None, "view_configuration_shopproduct_saved", True)
     for field in visible_fields:
-        configuration.set(None, "view_configuration_product_%s" % field, {"active": True, "ordering": 999})
+        configuration.set(None, "view_configuration_shopproduct_%s" % field, {"active": True, "ordering": 999})
 
     listview = ProductListView()
     column_names = [c.id for c in sorted(listview.columns, key=lambda x: x.id)]
     assert len(listview.columns) == len(visible_fields)
     assert column_names == visible_fields
-